+++ src/main/kotlin/com/cormontia/adventOfCode/year2023/utils/Coor.kt
@@ -1,3 +1,3 @@
 package com.cormontia.adventOfCode.year2023.utils
 
-data class Coor(val row: Int, val col: Int)
+data class Coor(val row: Long, val col: Long)
